style(workflow): Bump version in Test workflow
Now that the Commitizen action no longer treats having no work to do as
failure, there is no longer a need to cordon it off in a separate
workflow that does not notify on failure. Hence, move the call to the
Commitizen action to the Test workflow so that it only runs if all
pre-commit hooks pass. This guards against buggy releases. Uninstall the
pre-commit hooks installed by the official pre-commit action at the end
of the ScribeMD pre-commit-action to avoid preventing Commitizen or any
other subsequent action from performing Git commands. This is necessary
because, while the ScribeMD pre-commit-action adds pre-commit to the
PATH, this change does not persist after the action ends, but the Git
hooks are not similarly automatically cleaned up. Since this is a
breaking change, introduce a new no_bump input to allow the old behavior
to be retained by caller workflows that don't use Commitizen.