feat(action): Add bump input
Move the commitizen-action call from the Test workflow to the composite
action so that it can be used by repositories that only use the action
and not the workflow. Rename the no-bump input to bump, and invert its
semantics so that it matches the new bump input of the action and simply
ferries the input. Accept any compression in markdown-link-check so that
it doesn't receive a 400 when checking the docs.github.com link added to
the README.