docs(README): Lighten badge label backgrounds
Replace the previous default badge label background color of #555555
with the slightly lighter #666666. Configure https://coolors.co/ with
the logo color as the primary color, label background color (#666666) as
the secondary color, and text color (#FFFFFF) as the quaternary color.
Use the first recommended tertiary color for the message background
color. Use #343B42 as the label background color to match the workflow
status badge for badges on the same line.