docs(README): Document need to pass fetch-depth: 0
Document that pre-commit-action runs git remote set-head origin --auto
to support the new commitizen-branch hook and that, consequently, users
of pre-commit-action must pass fetch-depth: 0 to actions/checkout
regardless of whether bump is true. Previously, it was possible to run
pre-commit-action on a shallow clone (fetch-depth: 1) when bump was
false, but now a full clone (fetch-depth: 0) is always required.