refactor(action): Remove unneeded hook uninstall
We previously uninstalled pre-commit hooks prior to running the
Commitizen action to prevent them from blocking the auto-generated
commit. This was necessary because we used pre-commit/action, which
installed the pre-commit hooks prior to running them. We have since
ported off this deprecated action, so we no longer install the hooks.