fix(action): Add Ruby version to pre-commit key
Fix a bug related to partial restoration of the pre-commit cache.
Contrary to pre-commit's official documentation, Ruby pre-commit hooks
rely on the system version of Ruby. Ruby pre-commit hooks fail when the
version of Ruby installed on the runner doesn't match the version that
was present when the cache was created. The Ruby version was already
factored into the hash of **/.tool-versions in the cache key, but
specify it explicitly in both the key and restore key so that pre-commit
caches are no longer restored across Ruby versions. This change to the
cache key format invalidates all pre-commit caches. Introduce new
python-string, ruby-string, etc. outputs to the tool-versions step to
simplify building cache keys.